+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 ===================
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where registering a user with a social account that was used for a soft-deleted user was throwing an error, preventing completion of the registration process. ([#6](https://github.com/dukt/social/issues/6))
+
 ## 2.0.0-beta.16 - 2019-03-03
 
 ### Changed